requirements.txt
absl-py==2.1.0
aiohttp==3.9.4
aiosignal==1.3.1
antlr4-python3-runtime==4.9.3
asttokens==2.4.1
async-timeout==4.0.3
attrs==23.2.0
backcall==0.2.0
beautifulsoup4==4.12.3
black==23.10.1
bleach==6.1.0
cachetools==5.3.2
certifi==2022.12.7
cffi==1.16.0
charset-normalizer==2.1.1
click==8.1.7
colorama==0.4.6
comm==0.2.1
contourpy==1.1.1
crypten==0.4.1
cryptography==42.0.5
cycler==0.12.1
datasets==2.18.0
debugpy==1.8.0
decorator==5.1.1
dill==0.3.8
docker-pycreds==0.4.0
efficientnet-pytorch==0.7.1
einops==0.7.0
executing==2.0.1
filelock==3.9.0
flwr==1.8.0
flwr-datasets==0.1.0
fonttools==4.45.1
frozenlist==1.4.1
fsspec==2023.10.0
future==0.18.3
gitdb==4.0.11
GitPython==3.1.42
google==2.0.3
google-auth==2.27.0
google-auth-oauthlib==1.0.0
grpcio==1.62.1
huggingface-hub==0.20.2
idna==3.4
imageio==2.34.0
imageio-ffmpeg==0.4.3
importlib-metadata==7.0.1
importlib-resources==6.1.1
ipykernel==6.29.0
ipython==8.12.3
iterators==0.0.2
jedi==0.19.1
Jinja2==3.1.2
joblib==1.3.2
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
jupyter_client==8.6.0
jupyter_core==5.7.1
kaggle==1.6.6
kiwisolver==1.4.5
lazy_loader==0.3
lightning-utilities==0.10.1
linformer==0.2.1
Markdown==3.5.2
markdown-it-py==3.0.0
MarkupSafe==2.1.2
matplotlib==3.7.4
matplotlib-inline==0.1.6
mdurl==0.1.2
mpmath==1.3.0
msgpack==1.0.8
multidict==6.0.5
multiprocess==0.70.16
mypy-extensions==1.0.0
nest-asyncio==1.6.0
networkx==3.0
numpy==1.24.4
oauthlib==3.2.2
omegaconf==2.3.0
onnx==1.15.0
opencv-python==4.8.1.78
packaging==23.2
pandas==2.0.3
parso==0.8.3
pathspec==0.11.2
pathtools==0.1.2
pexpect==4.9.0
pickleshare==0.7.5
Pillow==9.3.0
piq==0.8.0
pkgutil_resolve_name==1.3.10
platformdirs==3.11.0
promise==2.3
prompt-toolkit==3.0.43
protobuf==3.20.3
psutil==5.9.8
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==15.0.2
pyarrow-hotfix==0.6
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.22
pycryptodome==3.20.0
pydantic==1.10.15
Pygments==2.17.2
pyparsing==3.1.1
pyspng==0.1.0
python-dateutil==2.8.2
python-slugify==8.0.4
pytz==2023.3.post1
PyWavelets==1.4.1
PyYAML==6.0.1
pyzmq==25.1.2
ray==2.6.3
referencing==0.34.0
regex==2023.12.25
requests==2.28.1
requests-oauthlib==1.3.1
rich==13.7.1
rpds-py==0.18.0
rsa==4.9
safetensors==0.4.0
scikit-image==0.21.0
scikit-learn==1.3.2
scipy==1.10.1
seaborn==0.13.2
sentry-sdk==1.41.0
setproctitle==1.3.3
shellingham==1.5.4
shortuuid==1.0.13
six==1.16.0
sklearn==0.0.post12
smmap==5.0.1
soupsieve==2.5
speedtest-cli==2.1.3
stack-data==0.6.3
sympy==1.12
tensorboard-data-server==0.7.2
text-unidecode==1.3
threadpoolctl==3.2.0
tifffile==2023.7.10
timm==0.9.8
tokenizers==0.15.0
tomli==2.0.1
torch==2.1.0+cu118
torcheval==0.0.7
torchmetrics==1.3.0.post0
torchvision==0.16.0+cu118
tornado==6.4
tqdm==4.65.0
traitlets==5.14.1
transformers==4.36.2
triton==2.1.0
typer==0.9.4
typing_extensions==4.4.0
tzdata==2023.3
urllib3==1.26.13
wandb==0.12.17
wcwidth==0.2.13
webencodings==0.5.1
Werkzeug==3.0.1
xxhash==3.4.1
yarl==1.9.4
zipp==3.17.0
